Update accounting logic for total encoded blob size  (#421)

+++ encoding/data.go
@@ -40,9 +40,9 @@ func (f *Frame) Length() int {
 	return len(f.Coeffs)
 }
 
-// Returns the size of chunk in bytes.
-func (f *Frame) Size() int {
-	return f.Length() * BYTES_PER_COEFFICIENT
+// Size return the size of chunks in bytes.
+func (f *Frame) Size() uint64 {
+	return uint64(f.Length() * BYTES_PER_COEFFICIENT)
 }
 
 // Sample is a chunk with associated metadata used by the Universal Batch Verifier
